requirements.txt
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0

loguru~=0.6.0
requests~=2.28.2